Move top-bar cart into own component
