+++ src/components/MobileHeader.tsx
@@ -1,9 +1,7 @@
-import IconCart from '@/assets/components/IconCart'
 import avatar from '@/assets/image-avatar.png'
 import mobileMenuIcon from '@/assets/svgs/icon-menu.svg'
 import logo from '@/assets/svgs/logo.svg'
-import { useAppState } from '@/data/globalState'
-import MobileCartDialog from './MobileCartDialog'
+import TopBarCart from './TopBarCart'
 
 /**
  * Renders the mobile header component.
@@ -13,10 +11,6 @@ import MobileCartDialog from './MobileCartDialog'
  * in the cart.
  */
 function MobileHeader() {
-  const quantityInCart = useAppState(
-    (state) => state.cartContents?.quantity ?? 0
-  )
-
   return (
     <header
       id="top-bar"
@@ -35,21 +29,7 @@ function MobileHeader() {
       </div>
 
       <div className="flex flex-row items-baseline gap-5">
-        {/* Cart button */}
-        <MobileCartDialog>
-          <button aria-label="Cart" className="relative">
-            {quantityInCart > 0 && (
-              <span
-                className="absolute right-0 top-0 -translate-y-1/2 translate-x-1/2 rounded-[6.5px] bg-sunshine-fg px-[7px] py-0 text-[10px] font-bold text-white"
-                aria-label={`${quantityInCart} items in cart`}
-              >
-                {quantityInCart}
-              </span>
-            )}
-
-            <IconCart fill="#69707D" />
-          </button>
-        </MobileCartDialog>
+        <TopBarCart />
 
         {/* Avatar/Your Account */}
         <img src={avatar} alt="Your account" className="h-6 w-6" />
